왜 삭제되는거야야야ㅑ양

+++ Watcha/src/service/review/DeleteReviewService.java
@@ -3,10 +3,13 @@ package service.review;
 import java.sql.Connection;
 import java.sql.SQLException;
 
+import Exception.LoginFailException;
 import Exception.PermissionDeniedException;
 import Exception.ReviewNotFoundException;
+import dao.MemberDAO;
 import dao.WatchaReviewDAO;
 import jdbc.ConnectionProvider;
+import model.Member;
 import model.WatchaReview;
 
 public class DeleteReviewService {
@@ -17,15 +20,21 @@ public class DeleteReviewService {
 		return instance;
 	}
 	
-	public void delete(int reviewId, int memberId) {
+	public void delete(int reviewId, int memberId, String password) {
 		System.out.println("서비스로 들어옴");
 		WatchaReviewDAO reviewDAO =WatchaReviewDAO.getInstance( );
+		
+		MemberDAO memberDAO = MemberDAO.getInstance();
 		
 		try(Connection conn = ConnectionProvider.getConnection( )){
 			try {
 				conn.setAutoCommit(false);
 				
 				WatchaReview review = reviewDAO.selectById(conn, reviewId);
+				
+				// userId를 이용하여 Member 객체를 select 해옴
+				Member member = memberDAO.selectMember(conn, memberId);
+				
 				// 게시글이 있는지 확인
 				if(review == null) {
 					System.out.println("오류 발생, 삭제할 리뷰 없어");
@@ -36,6 +45,12 @@ public class DeleteReviewService {
 					System.out.println("오류 발생, 삭제하는 사람이 작성한 사람이 아니야");
 					throw new PermissionDeniedException("사용자 권한이 없음");
 				}
+				
+				// 비밀번호 확인
+				if(!member.matchPassword(password)) {
+					throw new LoginFailException("비밀번호가 일치하지 않음");
+				}
+				
 				reviewDAO.delete(conn, reviewId);
 				// reviewDAO 를 이용해서 게시글 삭제 메소드를 실행
 				conn.commit( );
